Add Event struct with basic field and callback support
In all collection scenarios that we are aware of we need an event to
track the id, name, and format of the data. The Event struct
encapsulates the id, name, and format together into a single struct that
also has a callback closure. This allows other code to create and setup
an event, then define what to do when it is collected in a session.

We do not yet have a session object, however, the test and bench
simulate what a session would do. This ensures we have the right
encapsulation for later levels and establishes a baseline for basic
performance of the Event callback closure calls.