Introduce DataField/Ref and full data access
Most places we read events from aren't simply just the event payload.
They typically have an envelope around the event containing extra data,
such as the time of the event, the callchain, the TID or PID, etc.

The DataField struct represents where in the envelope/full data a given
piece of data resides at that point in time. DataFieldRefs use a Rc +
Cell to safely and panic-free share the locations of this data as
entries are read. These locations differ from each entry depending on
options being used, such as variable data sizes before a statically
sized entry.

Typical envelope from perf_events:
time: u64, // Static
call_chain: [u64], // Variable
event_data: [u8], // Variable
...

If in a given callback we want to get the time or call_chain we would
ask for a DataFieldRef of this items. As the event is parsed the start
and end of the full data is updated.

This allows two things that are important:
1. A data exporter can egress/copy any parts of the enveloped data
   without having to re-parse it to discover where things are. They also
   have the option to pull out only some data and egress it, or even
   parse/convert it to suite their egress needs.

2. During event callbacks, the DataFieldRef can be copied into the
   closure and used to get extra data that isn't available within the
   event_data directly. An example of this is the time the event fired.

To allow for the above cases add a new full_data slice to the event
callback signature which allows DataFieldRefs to be used to get this
data in future structs (like perf_events sessions).