Introduce PerfSession to parse perf_event formatted data
The first data we will collect resides within perf_event ring buffers.
The PerfSession struct owns all the contracts for how to parse the
entries returned from a perf_event ring buffer. However, it does not own
how that data is obtained.

The new PerfDataSource trait describes how to get data that is in the
perf_event format. This data could come directly from a streaming
network socket, a file, or in-memory ring buffers via the perf_event_open()
syscall.

This change only adds a single PerfDataSource for mock data when
testing. In future changes a PerfDataSource will be written for
in-memory ring buffers via perf_event_open(). Later we may have more as
needed. The idea is to split up the work of consuming data from a source
and the data record formats of perf_events.

The PerfSession exposes a DataFieldRef for callers for all entries that
it currently supports that a consumer *COULD* enable and that we
currently *SUPPORT*. We currently support up to PERF_SAMPLE_RAW via
PERF_RECORD_SAMPLE records. That means we are missing raw stack data
and user regiters (for DWARF unwinding features) and some other lesser
used features (such as LBR branches and CGROUP entries).

We also only currently handle PERF_RECORD_SAMPLE record types. Other
records will be added in future changes.