Introduce SharedData (Writable/ReadOnly) struct
In some cases, especially with perf_events, there is ancillary data that
isn't directly within the raw data. For example, each ring buffer is
associate with a CPU. If a caller needs to know the CPU and it's not
within the raw data, they are out of luck.

Provide a SharedData struct that allows a way for code parsing events to
get a readonly view of this ancillary data related to the event. This
allows a data source provider to provide this extra context without
requiring copying the data into another buffer or other unfortunate
workarounds to this problem.

Future changes will utilize these to accomplish ancillary data cases.