Introduce AncillaryData for perf_events
While parsing events we need to know the CPU it came from as well as
other things, like the sample type or read format of data to properly
decode some events.

Add AncillaryData struct to describe this type of data. Expose it out as
a ReadOnly data type, which allows it to be updated within the session
with outstanding references to it for event callbacks.

Callers get a readonly copy of the AncillaryData via the new
ancillary_data() function in the PerfSession struct.

Typical way to use this:
    let mut session = ...;
    let event = ...;

    let ancillary = session.ancillary_data();

    event.set_callback(move |full_data,_,_| {
        let mut cpu: u32 = 0;

        ancillary.read(|ancillary| {
            cpu = ancillary.cpu();
        });

        println!("Event was on CPU{}", cpu);
    });