Introduce Ruwind crate and proc/mod tracking
For proper DWARF unwinding we need to track the process lifetime and the
modules within it. We also need a way to parse and evaluate unwinding
directions.

Add the ruwind crate that lets us unwind using DWARF instructions. Add a
callstack helper that abstracts out the complexity of tracking all of
this. To start with this will just track process and module lifetimes.
Later changes will expose a clean way to unwind callchain data using raw
stack data from sample events.