Change RingBufOptions to a trait exposed by builders
Helpers need to be able to modify the low level options of the ring
buffer. Currently this cannot be done easily, as the RingBufOptions is a
struct that is passed into the builder once and then is never accessed
again. We also have some options that are just not allowed on some types
of RingBufferBuilders, such as the kernel events never actually
sampling.

Change RingBufOptions over to a trait that is only exposed out for
builder types that support those options. This allows helpers and other
parts of the program control these low level options and only when they
are actually available.