Introduce Error Reporting for Event Callbacks
Currently event callbacks cannot report any errors they encounter. We
want to be able to track when errors occur and from which events during
processing.

Modify the event callback signature to return a Result<(), anyhow::Error>.
Pull in the anyhow crate, which lets the event callbacks use any error
they wish, instead of relying upon a single one. This makes the
callbacks trivial to write with error handling as the ? operator can now
be used for any method that returns a Result<T, E> regardless of the E
type.

Update Event.process() to take in a Vec<anyhow::Error> to store errors.
Add a self test for event callbacks to ensures errors are propogating
out properly to the Vec. Errors are stored accumulative, so programs can
either accumulate all of the errors or process/log them as they arrive.
The PerfSession logs to the standard error by default, callers can
override this by calling PerfSession.set_event_error_callback() to do
something more custom.

Update all callers of Event.process() and Event.add_callback() to do the
right thing now. This means our callbacks no longer unwrap() and cause a
panic nor do they assume a default value for invalid data. In all cases
an error is returned if the data is not in the intended format or size.