Integration test running in a Docker container.
Implement a docker-based test that runs the debug cli for 5 seconds
and emits the individual events to the console.  At this time, the
test does not validate the output of the logger, but ensures that
the logger is properly setup and that there aren't any failures
that cause the process to panic or crash.