Add uprobe reg/unreg to tracefs
Some scenarios require creating on-demand probes for processes. We can
use various mechanisms for this. The most popular one being uprobes.

Add a uprobe and uretprobe register/unregister mechanism that will allow
creation of uprobes with a set of parameters. The register function
returns an event that fully describes the format. This allows callers to
register a probe, hook a callback, read the data for a period of time,
then finally unregister the probe easily.