Add enumerating function symbols in ELF
Various scenarios require getting symbols for functions. Typically this
is used while profiling the machine if a specific file format requires
it. It is also used when uprobes or breakpoints are to be used during
investigations.

Add get_symbols() method for ELF files into ruwind. This method takes
the burden to properly parse the start, end, and name of symbols which
are functions. The caller is given back each symbol. The caller is
responsible for any demangling if needed.