Add uprobe helpers
Scenarios that utilize uprobes need to be able to enumerate running
processes on the system, find which modules they are using, and see what
probes are available. We currently don't expose this, although we have
many of this already written.

Add a uprobes helper that exposes some of our internal functions
in a safe/expandable way. This allows tools to be built using uprobe
workflows easily.