Add uprobe example for allocations
Add a simple example showing how to find uprobes for a given process and
hookup reading when those probes run.

The current example finds a libc.so in the target process, finds the
malloc function, then adds a probe to this method. It then tracks the
memory usage on a per-second basis and prints to the console.