Exporting: Add Process FS access
Processes can be started in a different root or mount namespace. When
profiling or tracing tools need to get data a process creates, it must
do so under the same root or mount space.

Add a open_file() method to ExportProcess that uses a cached FD of the
/proc/pid/root directory and openat() libc call to open files. This
allows, for example, the /tmp/perf-pid.map file to be read after
collecting a profile for dynamic symbols.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>