Exporting: Abstract event details from callers
Callers wanting to export events should do so mainly via the
ExportSettings. They should not have to understand the flags and or
records required in most cases. They should only have to know high level
details like "I want context switches and CPU profiling".

Add a trait for the RingBufSessionBuilder to abstract out the events to
pull in based on the export settings.

Add a trait for the PerfSession to abstract out the correct way to
build and hookup an exporter to the session.

These significantly lead to simpler coding patterns and less chances of
someone doing something wrong now and in the future.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>