Event: Allow flags on events to indicate IP only
We'll often have sessions that mix profile-like events with low latency
events. These low latency events that have high volumes might not need
a callstack.

Add a flag to the Event struct. Introduce a no-callstack flag, which
causes just that event to get an IP only callstack. This allows sessions
to mix and match the verbosity of event callsites.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>