Callstacks: Handle IP only callstacks
There are two cases now when you want IP only callstacks. This can be a
global session level, because you only the IPs involved for your
analysis. Or it can be you normally want callstacks, but some high
volume event doesn't need them or it's too costly to collect (especially
with DWARF).

Add an ip_only() extension to CallstackHelper to turn all events into
only ip samples. Handle the proper sample_type changes to achieve this.

When reading frames, if callchain/callstack data is empty, use the
ip_field from the session to try to get the IP instead.

This also aligns the CallstackHelper to handle adding the appropriate
sample type to each enabled event. Before only DWARF cases would get
this logic, and it was confusing to know who should set the flags. This
change makes the CallstackHelper consistently the one to setup the flags
now that we have an ip_only escape hatch for all events or specific
events.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>