Exporting: Introduce ExportGraph and PerfView format
We want to support a wide range of formats for profiling and
collections. A lot of these formats rely upon mapping virtual addresses
to relative virtual addresses of modules. We want to prevent formats
from having to re-do this work over and over.

Add ExportGraph struct that takes an ExportMachine and ExportProcess and
generates a graph. The graph has nodes that point to a common module the
code is from and other details. Samples are deduplicated and aggregated
together.

Add a format module and introduce PerfView XML output from an
ExportGraph as a trait extension (to_perf_view_xml). This sets up a
pattern for process-level formats to utilize for exporting with minimal
effort.

Several processes can be combined into a single ExportGraph. However, it
should be clear that process isolation is not inserted at the graph
level. Simple things, like combining many processes with the same comm
name are easily handled. Tools can choose to export at a common level
(such as comm name) or at a per-PID level, depending on what they need.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>