docs: Add docs for Events and Sharing
We need to ensure our contributors and consumers are able to properly
use our crate. To do so, we need to have some documentation and graphs
that clearly show how we expect things to be put together (and why we
have done it like this).

Add EVENTS.md that covers the aspects of what an Event is and how data
makes it way from a pipeline into the Event closures. Include calling
out how a closure can get a specific field of an event.

Add SHARING.md that covers the aspects of sharing data between
pipelines, events, and closures to make things easier.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>