Properly set the end address in KernelSymbolReader.
KernelSymbolReader reads one line of /proc/kallsyms at a time.
Each line contains the start address, type of symbol, and the symbol
itself.
The end address is the start address of the next line minus one.
In cases where the next line is not a symbol for code in the text
section, rather than skip the line, we should save the end address prior
to skipping the line.