one_collect: Introduce OpenAt
Rust does not currently provide a way to open a file relative to another
file. This can be done via the libc::open_at() syscall. We need to be
able to do this to ensure we can cache process root directories in the
event they exit, but we still need to access files from their namespace.

Add an OpenAt struct that wraps the libc::open_at() syscall. Use this in
place of bespoke logic in the ExportProcess struct. In future structs we
will use this as well.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>