procfs: Get the NSpid for a process
Processes on Linux have a PID from the collecting namespace and the PID
namespace that process belongs to. In order to get files that process
creates with it's own PID in a namespace, the namespace specific PID
must be used (NSpid).

Add a method to read the NSpid from the /proc/<pid>/status file. This
gives a collector both PIDs, which is required for perf_maps and other
things like diagnostic sockets.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>