perf_event: Add drop event to PerfSession
Sometimes an extension/helper would like to run some actions when the
PerfSession gets dropped by Rust. An example of this is a processing
thread that a helper spawned should stop and cleanup when the
PerfSession completely stops (drops).

Add a drop_event, like any other PerfSession event. Call this event when
the PerfSession drops by Rust.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>