Helpers: Introduce DotNet Helper
When you are typically collecting profiling data for DotNet, you need to
enable some features. This is typically perf-map file generation so
DotNet applications will output dynamic symbols into the tmp directory.

Add a DotNetHelper struct and matching trait to the PerfSession. Add an
option to create perf-map files dynamically via the EventPipe IPC
mechanism (see Link). Use a worker thread to enable these, as it can be
slow and we don't want to block the main event loop. We also only want
to cause a symbol rundown at a process at a time. It's likely bad to
cause many to occur in parallel on a production environment.

Allow callers to delete the perf-map files that we caused to be
generated. If existing perf-map files were there due to environment
variables, then we leave those alone.

Link: https://github.com/dotnet/diagnostics/blob/main/documentation/design-docs/ipc-protocol.md#enableperfmap
Link: https://learn.microsoft.com/en-us/dotnet/core/runtime-config/debugging-profiling#export-perf-maps-and-jit-dumps
Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>