ruwind: Read notes for Elf files
Typically we need to read elf note sections. These are extensions to elf
files for various scenarios. We need these typically for build-id
calculations and version data.

Add get_notes() to Elf file and expose a Note struct that enables
reading the name and data of the note (as binary and as string).

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>