Exporting: Reduce PathBuf allocations
The procfs struct hides that it allocates a PathBuf during common
operatoins, such as getting the NSpid. These paths need to be fast, so
allocations are an issue.

Make common procfs operations take a mut PathBuf ref instead of
allocating one. Then the caller is free to decide if they want to use a
shared or new PathBuf.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>