Ruwind: Abstract out Process/Module
Ruwind currently has a concrete Process and Module struct which it uses
to track details for unwinding code. This enforces a model where Ruwind
owns tracking the full details of the machine.

When Ruwind is used in other places where this tracking is already
occurring, it results in double the memory overheads. Instead of
tracking this separately, it would be ideal to allow either using the
concrete structs or another compatible struct from another system.

Add Unwindable and CodeSection traits to Ruwind to describe the required
attributes in order to unwind. Add implementations for Unwindable to the
Process struct and CodeSection to the Module struct.

Update unit tests, etc.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>