Exporting: Enable external lookup for callstacks
Ruwind now offers a way to support external lookup of modules and
processes. The ExportMachine already has these stored that we can use.

Enable a pattern for callers to use that ensures the ExportMachine can
flip on external lookups and also do a custom unwind using it's own
process lookup tables.

Callers now pass a CallstackHelper struct to the ExportSettings struct.
The ExportMachine then takes cares of wiring things up in the correct
order instead of requiring the caller to know that. This makes it much
safer on our side to refactor later if required without breaking things.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>