Exporting: Fix existing node merge values
The merge algorithm in our ExportGraph uses a lookup to find the
existing parent node to merge from. The existing node case, the value of
the sample wasn't being charged down the node graph toward the root.

Add a charge() method to add samples values from the parent down to the
root to properly record sample inclusive values.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>