Exporting: Fix kernel address lookup and bad paths
While testing the unwinder I found some cases where the perf_export
would fail to export. This happened when the comm name was a kworker
with a / and : in the name. I also found that kernel symbol reset wasn't
properly reseting the next_ip and current_ip which resulted in missing
symbols. When the symbols were missed it also showed that the address of
kernel IPs were being saved incorrectly.

Replace / and : during export with _.
Reset current_ip and next_ip in KernelSymbolReader.
Don't use RVA calculations on Kernel addresses.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>