BPF: Allow attaching to eBPF perf_event maps
eBPF programs that utilize perf_events to output data can be attached by
setting the perf_events map array to perf_event FDs. The PerfSession
already exposes the perf_event FDs, however, we need a way to set an
eBPF map to those values in order for it to work.

Add bpf module that abstracts out the bpf syscall from the PerfSession.
Add methods to attach the perf_fds to a map either by ID, Path, or FD.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>