BPF: Allow an event per-BPF map
When perf_events are attached to BPF maps the events will go out to the
RingBuffers. However, we don't have a way for callers to get this data
out cleanly.

Add an Event parameter when attaching to BPF maps. Store the event in
HashMap that aligns to the buffer ID that was attached to the map.
Lookup the Event based on the buffer ID and process the data via the
event callbacks.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>