BPF: Ensure BPF events can disable callstacks
When mixing BPF events with other BPF programs, you may have some
programs that don't need callstack data, while others do.

Update the create_bpf_files() trait method to take in an optional Event
that can describe no callstacks via the Event flags.

Use the passed in Event when adding a shared bpf map to describe if
callstacks are wanted or not, just as we do for tracepoint events.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>