Perf: Allow multiple target_pids
Some tools need the ability to only enable ring buffers for a set of
PIDs. Today we only allow a single PID, which works for some cases, but
not all.

Save all pids that are passed via the RingBufSessionBuilder::with_target_pid().
This changes behavior where before this call would overwrite. Now it
will accumulate all the PIDs that are specified and do the right thing
for the caller.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>