Perf: Honor target_pids() during state capture
When tools are using target pids, they do not expect to see mmap and
comm records from external pids. Currently we scan the entire
/proc/*/ files, regardless of this setting.

Add target_pids() to PerfDataSource trait to allow for sources to
describe any target PIDs (if any).

Utilize target_pids() in PerfSession when asked to capture the
environmnet. Comms and mmaps filter out to only the requested PIDs if
they were returned from the PerfDataSource.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>