Event: Add EventData for callbacks
Currently Event has a callback signature that is passed the event_data,
format and full_data. If we need to add more, that would require all
callers to change their closure signatures.

Make the callback signature take a single parameter of an EventData
reference. This way, we can expand this over time without breaking
existing users. This will cause a one-time breaking change, but there is
minimal users, so it's better to do so now than later.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>