Symbols: Allow external file for Kernel Symbols
Currently the KernelSymbolReader only reads from /proc/kallsyms, which
is appropriate for most cases. However, sometimes you need to read this
data from another location. An example of this is when it is saved
within a file format not on the original machine.

Add set_file() method to KernelSymbolReader which overrides the default
file to use when parsing kernel symbols.

Update namespacing for File and BufReader.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>