Symbols: Fix kernel end address of methods
The end address of a kernel method within kallsyms only used the next
method's start address. This works fine for symbol resolving, but
doesn't show the function size properly. We should be accurate with the
start and end of the method to ensure scenarios beyond just symbol
resolution work properly.

Always calculate the end address by using the next line entry address,
regardless of it's type. This is done only for the first next line
found, but not subsequent lines of methods.

Add unit test and kernel symbol map assets to ensure the logic is
correct.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>