Symbols: Handle Kernel Symbol edge-cases better
There are two edge cases for kernel symbols that we need to deal with.
The first is that kernel symbols sometimes do not come directly from the
device and are stored elsewhere. The second is that kernel symbols are
sometimes from modules other than the kernel itself, and we want the
symbol name to conform to this that aligns with other tools.

Add add_kernel_mappings_with() which allows passing any symbol reader to
it to resolve the kernel symbols with. The existing
add_kernel_mappings() simply leverages this now with a default
KernelSymbolsReader.

Change kernel symbols that are from a sub-module to the existing
standard "[module] method" format, instead of "module!method" format.
Update unit test to ensure this is accurate. Do not export "vmlinux" as
a module, since that is already described in the module section it
originates from. Before we would get "vmlinux!module!method", we now get
"vmlinux![module] method". When no module is used, we get
"vmlinux!method".

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>