PPROF: Initial support for PPROF format
The PPROF format is well documented and supported by various tools. It
also is the leading format for the emerging OTEL Profile Format.

Add support for PPROF via the to_pprof and to_pprof_file traits. This
allows either a gzip file to be output by use of pprof native tooling or
the raw uncompressed protobuf bytes for use in the emerging OTEL Profile
Format.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>