ETW: Initial drop of EtwSession
On Windows we need to be able to get events and profiling data. This is
typically done via ETW.

Add an EtwSession that can be used to add events via providers, levels,
and keywords. As events are read, closures are invoked, the same as we
do on Linux perf_events.

Signed-off-by: Beau Belgrave (HE/HIM) <beaub@linux.microsoft.com>