ETW: Add Event Hooks and Processing
ETW has some interesting traits around multi-plexing older system events
into other sessions. We need to be able to add these newer sessions that
allow up to 8 sessions at a time, instead of 1, to be able to receive
events. This means we need to register with one provider ID, but then
process the events using the non-multiplexed provider ID, which is
different.

We also need to be able to enable various properties for events, such as
stacktraces. The system events use a different method for this than all
the rest of the events. Add methods that can indicate when to use the
actual filtering and properties.

Add proper lookups for events that handle the multiplexed scenarios as
well as the normal scenarios equally.

Add hooks for pipelines to run code when the session is starting,
started, stopping, or stopped. This allows various scenarios to work
well, such as ensuring full machine state has been captured before
starting to profile the machine.

Signed-off-by: Beau Belgrave (HE/HIM) <beaub@linux.microsoft.com>