Event: Add per-OS extension for Event
ETW on Windows needs to track the Provider per-Event so that it can be
added to lookups later. On Linux only the ID is required, however,
Windows needs a Guid and the ID to properly look things up. The level
and keyword are also used to enablement, which should be stored for
dynamic enablements of providers.

Add an empty EventExtension struct for Linux.
Add a Provider, Level, and Keyword EventExtension struct for Windows.

Signed-off-by: Beau Belgrave (HE/HIM) <beaub@linux.microsoft.com>