ETW: Add Kernel Callstacks
The kernel/system providers have a special way to enable callstack
collection within ETW. These enable via the TraceSetInformation() API
instead of the normal event properties, such as the
EVENT_ENABLE_PROPERTY_STACK_TRACE property on ENABLE_TRACE_PARAMETERS.

Add add_kernel_callstack() to EtwSession to allow for callers to
describe this. Add enable_kernel_callstacks() to the ABI that will
invoke the TraceSetInformation() API. Call enable_kernel_callstacks()
when the session is started with any kernel events added.

Hide this complexity from callers for common kernel events, such as
profile, cswitch, ready_thread by taking the properties that would
normally get enabled for an event. If the properties is a callstack,
silently add the event to have callstacks for the caller.

Add cswitch, ready_thread, and callstack events to common session
events. The callstack event is special, it does not enable the provider
itself. However, it must be in the provider lookup to find these events
when the kernel logs them when they are enabled.

Signed-off-by: Beau Belgrave (HE/HIM) <beaub@linux.microsoft.com>