ETW: Async Callstack Pattern
ETW (and soon Linux) supports async callstack collection when a
callstack originates in the kernel. An example of this is the
SampledProfile event (1), if the sample is on a CPU that is running in
kernel mode, the kernel callstack is emitted after the SampledProfile
event as a StackWalk event (2). The tricky part, is if the process also
has user-mode code that lead up to the kernel call (very likely) then
yet another StackWalk event (2) is emitted. We don't want every event to
know this detail or handle if user callstacks fail or never come in.

To ensure this is done consistently, add mechanisms to register when a
full async callstack has been read (add_async_frames_callback) to the
Windows CallstackReader struct. For cases when a user callstack never
comes in, expose add_flushed_callback to the Windows CallstackReader
struct as well.

Expose a built_callback in EtwSession to ensure the CallstackHelper can
run a closure to see if kernel callstacks are enabled for anything. If
they are, then hookup the CallstackHelper to the EtwSession, utilizing these newly
exposed callbacks correctly.

Add self tests to ensure the lookup works correctly for all cases.

NOTE:
The Linux SFrame work (3) will likely also need a similar pattern, although
the way to lookup matches will be different. When that occurs I expect
the add_async_frames_callback and add_flushed_callback to be in both OS
modules.

1. https://learn.microsoft.com/en-us/windows/win32/etw/sampledprofile
2. https://learn.microsoft.com/en-us/windows/win32/etw/stackwalk-event
3. https://lwn.net/Articles/932209/

Signed-off-by: Beau Belgrave (HE/HIM) <beaub@linux.microsoft.com>