ETW: Hookup CPU Profiling
ETW supports CPU profiling the same as on Linux. There are a few
differences though. ETW's SampledProfile event does not log the PID, it
only logs the TID. The async callstack that comes in however, does store
the PID. This means that we always need to get the callstack event or we
need to track thread start/exit events. Typically profiling wants
callstacks, so we will correlate this at the time we get the async
callstack. For samples that had errors unwinding, we will store these in
an Unknown process to prevent skewing the machine wide utilization
numbers.

Add callbacks for async_frames and flushing. Add samples from both of
these callbacks to the appropriate process. In the case of flushing, we
will not have the PID. Add these to unknown process without a local PID
association to prevent machine wide utilization skew.

When the idle thread is the one getting a callstack, do not store it for
later unless the count is more than 1.

Signed-off-by: Beau Belgrave (HE/HIM) <beaub@linux.microsoft.com>