ETW: Wire up perf_export to ETW
In the perf_export example we wire it up to perf_events, but not ETW. We
need to show how to use both. During this, I found that the kernel
mappings need to also be mapped in a unique way when using ETW. This is
because the kernel mappings are saved in the PID 0 process. We need to
pull these mappings out to each process since Windows has several
mappings in the kernel vs a single mapping on Linux.

Update perf_export to use ETW cpu sampling.

Split add_kernel_mappings_with to OS half and common half.

Ensure the KernelSymbolReader is marked done on reset to ensure partial
reads aren't restarted when open fails.

Signed-off-by: Beau Belgrave (HE/HIM) <beaub@linux.microsoft.com>