Exporting: Add support for mapping load times
On Windows it's very common for modules to be loaded and unloaded within
a process. On Linux it's less common, but occurs with dlopen() and
dlclose(). When trying to determine which module a given sample is for
time needs to be taken into account. For example, if a module A loads
and then unloads, then module B loads into overlapping space, it's
ambigious as to if module A or B matches a given sample IP. However, if
module A's time and module B's time is stored, we can use the sample
time to determine which mapping it is. We can also find the most recent
mapping for live unwinding cases when a previous module was unloaded.

Add time properties to ExportMapping and related functions within the
exporting helper.

Add ExportMappingLookup and AddressLookup structs to ensure we can find
addresses with time efficiently and consistently. Don't use a fancy
lookup if there are only a few modules, etc.

Add self tests to ensure lookups work.

Fix ExportGraph to not assume Linux path separators and use the per-OS
separator value for file path trimming to file name.

Signed-off-by: Beau Belgrave (HE/HIM) <beaub@linux.microsoft.com>