Exporting: Universal support for all OS types
When tools are built that require cross platform support, we should have
a way to do that in a scalable way. Currently each OS type has a session
type and a builder type. We need a way to abstract this out, while still
allowing extension points by other helpers as well as out-of-tree
customizations that work for all OS types.

Add a Universal module under Exporting module.

Add UniversalExporter struct which can be used on all OS types for
getting an ExportMachine from after a period of time or event elapses.

Add extension points in UniversalExporter that other helpers (and
out-of-tree code) can utilize to do certain actions for the OS specific
Builder before parsing and allow modifying the ExportMachine before
being returned to the caller.

Later changes will include other helper support, like DotNet, to the
universal exporter to ensure we have a scalable way to enable
per-language support and other features across all the OS types.

Signed-off-by: Beau Belgrave (HE/HIM) <beaub@linux.microsoft.com>