Exporting: Fix perf_map resolution path
When paths are used for process specific files, they need to be
relative. The PathBuf should be cleared to ensure we don't have extra
components in the path and ensure it's going to be relative.

Clear the path_buf before using it, since it could have other path
components already within it from the previous call. Use /tmp, which
will be handled via the OpenAt logic to use a relative path while also
showing clearly that's the path we want without other components before
it.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>