Universal: Support per-CPU buffer sizes
Callers expect to be able to set per-CPU buffer sizes on all operating
systems we support. This is exposed directly in our Linux side, but is
not on the Windows side.

Add support on the UniversalExporter to set the per-CPU buffer size
regardless of OS.

Add support to set the per-CPU buffer size for ETW.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>