Universal: Add universal symbol trait
The details of collecting symbol metadata for later resolution and the
actual local parsing logic of symbols is unique per OS. Callers that
wish to run on all the supported operating systems need an abstraction,
so we can evolve this overtime.

Add the UniversalSymbol trait which allows capturing symbol metadata for
later resolution (either locally or remotely, file format depending).
Ensure the trait can also resolve local symbols for the kernel,
file-backed, and anonymous methods seen during the capture.

Implement the UniversalSymbol trait for the ExportMachine. This allows
callers to describe their flow, regardless of OS in a concise way
without many flags or closures ontop of the UniversalExporter struct.

For example, when we have a DotNet helper that works upon the
UniveralExporter, it can give back universal structs that keep virtual
symbols alive until they are resolved. This type of pattern, where
virtual/JIT'd symbols need to have a lifetime associated with them works
much better with an abstracted out trait vs a closure system. On Linux
the DotNet helper will enable perf_map file generation, but disable it
once the session is over. However, the resulting perf_map files are
quite large. They typically need to be deleted after they have been
used for resolution. This is done via the DotNet helper returning a
struct that can be used to clear the temporary state. We need to ensure
this type of pattern persists beyond per-OS flows and works well with
universal OS flows.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>