Universal: Implement Default for ExportSettings
In a univeral workflow we need to be able to define optimal default
settings for per-CPU buffer sizes when an unwinder is being used. This
logic must be done in a way that it can be overriden when necessary and
also allow for custom ExportSetting attributes, such as custom events.

Implement Default for ExportSettings so that we can enable DWARF
unwinding on x64 on Linux for callers without them having to know the
details. Each OS will establish it's own defaults.

Add a has_unwinder() method for both the CallstackHelper and
ExportSettings to allow the Universal module to increase per-CPU buffer
sizes when an unwinder is being used (currently only Linux supports
this).

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>