ETW: Align parse_until closure with perf_events
The parse_until method takes a closure that indicates when to stop. On
Linux this closure is run multiple times on the same thread. On Windows
this closure is run a single time on a different thread.

As part of the Universal work, it became clear that we need a single way
to write a closure that works on Windows and Linux the same way.

Update ETW closure pattern to require a bool return that indicates
whether to stop or not. Run the closure multiple times until true is
returned, sleeping a short period in between checks. Callers can simply
delay once or periodically query for status, as they can on Linux.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>